Changed site-wide aesthetic to encyclopedia style, focused site objective
There were too many dropdown menus on mobile, so I've refocused the site purpose, aesthetic, and navigation.  The purpose was always intended to be the encyclopedia, so the site is now focused on that.  Now there is only one navigation menu, which is combined with the encyclopedia navigation, and the entire site resembles the encyclopedia.